refactor(otp-input-kit): move tests into entry subpackage
